Issue #22: Sql type of autoNumber column
Changing the original ordering of TypesMap by sorting it alphabetically
changed program behavior of COUNTER (and possibly other) datatypes.
COUNTER should be mapped to INTEGER, however was first translated to INTEGER
then to SMALLINT due to reordering and the application of String.replaceAll
in construction of the sql string.

Thanks to @Takumi-Inoue-hulft